css反映

+++ app/Http/Controllers/Tweet/DeleteController.php
@@ -11,9 +11,12 @@ class DeleteController extends Controller
     /**
      * Handle the incoming request.
      */
-    public function __invoke(Request $request)
+    public function __invoke(Request $request, TweetService $tweetService)
     {
         $tweetId = (int) $request->route('tweetId');
+        if(!$tweetService->checkOwnTweet($request->user()->id, $tweetId)) {
+            throw new AccessDeniedHttpException();
+        }
         $tweed = Tweet::where('id', $tweetId)->firstOrFail();
         $tweed->delete();
         return redirect()->route('tweet.index')
